requirements.txt
fastapi
sqlalchemy
uvicorn
python-dotenv
bcrypt
python-jose
google-auth
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
pydantic
